pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gob"
version = "0.1.0"
description = "Homelab Intelligence Framework"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.19.0",
    "psutil>=5.9.0",
    "numpy>=1.24.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "tomli>=2.0.0",
    "openai>=1.12.0",
    "pydantic",
    "fire",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["core*", "mini*", "nano*", "main*", "heavy*", "zero*"]